Add support for labels
GitHub issues and pull requests can have labels assigned to them.
Filtering issues based on these labels enables categories of issues to
be analysed. This commit adds the ability to include columns in the csv
file for the labels to allow the data to be filtered based on the
presense or absense of specific labels.

Connected https://github.com/rcbops/u-suk-dev/issues/359