Fix date normalisation with non-date columns
The date normalisation function assumed that the data finishes with a
contiguous set of date columns. This commit modifies the function so
that it has fixed start and end points to allow additional information
to be added by extending the data with additional attribute columns.

Connected https://github.com/rcbops/u-suk-dev/issues/360