Add Pending-SHA-Update column
The Waffle board [1] includes the column 'Pending SHA update' and is
sited between 'Needs review' and 'Done'. This column is represented by
the label 'status-pending-sha-update' in GitHub.

This commit updates the output from gh2csv to include the date and time
when this label was added so that it is properly represented when
analysing the data.

[1] https://waffle.io/rcbops/u-suk-dev

Connected https://github.com/rcbops/u-suk-dev/issues/484